Updated FitsToCSV.py, started a development document (also slightly changed Image.py)
Completed the preliminary version of FitsToCSV.py, but it has yet to be tested on images.  To test it,
I will upload it to my account on the CENPA server so that I can access the DAMIC data.  In the
development document I started with stating the preliminary goals of the project, and my plan for
FitsToCSV.